+++ checks/path-dep-lib-reuse/print/src/lib.rs
@@ -1,0 +1,3 @@
+pub fn print(s: &str) {
+    println!("{}", s);
+}